# RepFabric.Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My API",
    "version": "v1"
  },

  "paths": {
    "/api/v1/excel/upload-template": {
      "post": {
        "tags": [
          "Excel"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "File"
                ],
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTemplateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/excel/fill-template": {
      "post": {
        "tags": [
          "Excel"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExcelMappingRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExcelMappingRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExcelMappingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.ms-excel.sheet.macroEnabled.12": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/vnd.ms-excel.sheet.macroEnabled.12": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/vnd.ms-excel.sheet.macroEnabled.12": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/excel/templates": {
      "get": {
        "tags": [
          "Excel"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ExcelMappingRequest": {
        "required": [
          "mappings",
          "templateFileName"
        ],
        "type": "object",
        "properties": {
          "templateFileName": {
            "minLength": 1,
            "type": "string"
          },
          "mappings": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Mapping"
            }
          }
        },
        "additionalProperties": false
      },
      "Mapping": {
        "required": [
          "cell",
          "fieldType"
        ],
        "type": "object",
        "properties": {
          "cell": {
            "minLength": 1,
            "type": "string"
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "fieldType": {
            "minLength": 1,
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UploadTemplateResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "storage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}